api/requirements.txt
fastapi==0.115.12
python-dotenv==1.1.0
bcrypt==4.3.0
PyJWT==2.10.1
uvicorn==0.34.2
